V0.1
User can now manually choose which executable to start. Make sure that this is a Runescape executable for now. After that, the user can launch this executable with the other button. After closing the executable, a messagebox will pop up with the right exitcode. This code is translated to a string with a dictionary.