V0.2
Added: Mouse click button. This will press at position 500, 500 on the executable screen. This is still hardcoded. 

Fixed:
Process start was freezing the form untill the process was closed. Fixed it by creating the process in a new thread.

To be fixed:
Thread does not wait untill process has been ended.